Skip unzip for already existing files

+++ sass-embedded-host/src/main/java/de/larsgrefer/sass/embedded/util/IOUtils.java
@@ -50,7 +50,9 @@ public class IOUtils {
             } else {
                 ensureDirectory(entryFile.getParentFile());
 
-                Files.copy(zipInputStream, entryPath, StandardCopyOption.REPLACE_EXISTING);
+                if (Files.notExists(entryPath) || !entryFile.exists()) {
+                    Files.copy(zipInputStream, entryPath, StandardCopyOption.REPLACE_EXISTING);
+                }
             }
             zipInputStream.closeEntry();
